+++ nomad/client.go
@@ -31,7 +31,7 @@ func IsConnectionErr(err error) bool {
 
 const (
 	DeploymentStatusRunning  = "running"
-	DeploymentStatusCanceled = "canceled"
+	DeploymentStatusCanceled = "cancelled"
 	DeploymentStatusSuccess  = "successful"
 	AllocationStateDead      = "dead"
 	Running                  = "running"
@@ -61,28 +61,44 @@ func NewClient(config *api.Config) (*Client, error) {
 	return &Client{API: api}, nil
 }
 
-// TODO maybe improve the logging?
 func (n *Client) waitForDeployment(ctx context.Context, jobID string) error {
-	ctx, cancel := context.WithTimeout(ctx, time.Minute*5)
+	ctx, cancel := context.WithCancel(ctx)
 	defer cancel()
+
+	ticker := time.NewTicker(time.Second * 30)
 
 	for {
 		select {
 		case <-ctx.Done():
-			return nil
+			return ctx.Err()
 		default:
 			time.Sleep(time.Second * 4)
+
+			job, _, err := n.API.Jobs().Info(jobID, &api.QueryOptions{})
+			if err != nil {
+				return err
+			}
+
 			deployments, _, err := n.API.Jobs().Deployments(jobID, true, &api.QueryOptions{})
 			if err != nil {
 				return err
 			}
 
+			timedOut, err := n.jobTimedOut(ctx, ticker, jobID)
+			if err != nil {
+				return fmt.Errorf("failed to tell of job timed out: %w", err)
+			}
+
+			if timedOut {
+				return fmt.Errorf("failed to run %s job: starting deadline has been exceeded", jobID)
+			}
+
 			for _, dep := range deployments {
-				log.Printf("deployment (%s) update for job: %q, status: %q, another: %s", dep.ID, dep.JobID, dep.Status, dep.StatusDescription)
+				log.Printf("Update for job: %q, jobStatus: %s, deploymentStatus: %q, another: %s", dep.JobID, *job.Status, dep.Status, dep.StatusDescription)
 
 				switch dep.Status {
 				case DeploymentStatusCanceled:
-					return fmt.Errorf("failed to run %s job", jobID)
+					return fmt.Errorf("failed to run %s job: %s", jobID, dep.StatusDescription)
 				case DeploymentStatusSuccess:
 					return nil
 				}
@@ -125,16 +141,17 @@ func (n *Client) RunAndWait(ctx context.Context, job *api.Job) error {
 	return nil
 }
 
-func (n *Client) Stop(ctx context.Context, jobID string, purge bool) (bool, error) {
+// Stop stops a specific job
+func (n *Client) Stop(ctx context.Context, jobID string, purge bool) error {
 	jobs := n.API.Jobs()
 
 	writeOpts := new(api.WriteOptions).WithContext(ctx)
 	jId, _, err := jobs.Deregister(jobID, purge, writeOpts)
 	if err != nil {
 		log.Printf("error stopping the job: %+v", err)
-		return false, err
+		return err
 	}
 
 	log.Printf("Stopped Job: %+v - %+v", jobID, jId)
-	return true, nil
+	return nil
 }