+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 - [211](https://github.com/vegaprotocol/vegacapsule/pull/211) Improve importing tendermint keys, re
 - [215](https://github.com/vegaprotocol/vegacapsule/pull/215) Removed old network parameter `market.monitor.price.updateFrequency`
 - [221](https://github.com/vegaprotocol/vegacapsule/pull/221) Downgrade to Tendermint v0.34
+- [217](https://github.com/vegaprotocol/vegacapsule/pull/217) Improve failure detection and add default logging sidecar to copy logs to home folder
 
 ### 🐛 Fixes
 - [167](https://github.com/vegaprotocol/vegacapsule/issues/167) Fix validators filter in tendermint generator